requirements_api.txt
torch==2.5.1
torchvision==0.20.1
timm==1.0.13
fastapi[all]==0.115.6
uvicorn==0.34.0
hydra-core==1.3.2
